Update insert user method

+++ AddressBook-backend/src/main/java/com/springboot/app/model/Register.java
@@ -1,6 +1,7 @@
 package com.springboot.app.model;
 
 import javax.persistence.Column;
+
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -11,9 +12,10 @@ import javax.persistence.Table;
 @Table(name ="register_tb")
 public class Register {
 	
+	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long id;
+	private int id;
 	
 	@Column(name ="name")
 	private String name;
@@ -47,10 +49,10 @@ public class Register {
 	}
 	
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
 	public String getName() {
